Add prometheus exporter page for hashrates
Adds a new http route /metrics for embedded http server. Prometheus
style metrics are getting displayed on this page for GPU hashrates,
producing two metrics:

* xmr_stak_amd{thread_id="x", interval="10/60/900"} 123
* xmr_stak_amd_total{interval="10/60/900"} 123

Doesn't decoreate other prometheus metrics which are usual in go based,
exporters.

Could be useful for hashrate monitoring with prometheus and grafana as
well as alerting if hashrates are going to zero. If hashrate is not
available 0.0 value is returned.